fix: remove fileblocks statistics
This patch accomplishes a couple of things. First, it removes the
fileblocks stats. These stats are an implementation detail and should
not be exposed to the user. Secondly, the queries are impossible to do
in cassandra.

This patch also fixes some Cassandra aggregation queries that are not
supported but were missed because our Cassandra driver is mocked with
sqlite (which does allow aggregation functions).